requirements.txt
alpaca-py
anthropic
apscheduler
requests
pandas
numpy
